pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Multiple projects ranging from LLM, Agents & Multi agent architecture to MCP and beyond!"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic==0.70.0",
    "autogen-agentchat==0.6.1",
    "bs4==0.0.2",
    "flask>=3.1.1",
    "gradio==6.0.0",
    "httpx==0.28.1",
    "ipywidgets==8.1.7",
    "langchain-anthropic==1.0.0",
    "langchain-community==0.3.31",
    "langchain-openai==1.0.3",
    "langgraph==1.0.3",
    "langgraph-checkpoint-sqlite==2.0.10",
    "langsmith==0.3.45",
    "lxml==5.4.0",
    "mcp-server-fetch==2025.1.17",
    "mcp[cli]==1.23.0",
    "openai-agents==0.6.1",
    "playwright==1.52.0",
    "plotly==6.1.2",
    "polygon-api-client>=1.15.3",
    "psutil==7.0.0",
    "pypdf==5.9.0",
    "pypdf2==3.0.1",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "requests==2.32.5",
    "sendgrid==6.12.3",
    "setuptools==80.9.0",
    "smithery==0.1.0",
    "speedtest-cli==2.1.3",
    "wikipedia==1.4.0",
    "gitpython==3.1.44",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.12",
    "aioice==0.10.1",
    "aiortc==1.13.0",
    "aiosignal==1.3.2",
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "autogen-core==0.6.1",
    "av==14.4.0",
    "azure-ai-agents==1.2.0b3",
    "azure-ai-projects==1.0.0b12",
    "azure-core==1.34.0",
    "azure-identity==1.23.0",
    "azure-storage-blob==12.25.1",
    "beautifulsoup4==4.13.4",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cloudevents==1.12.0",
    "colorama==0.4.6",
    "comm==0.2.2",
    "cryptography==45.0.4",
    "dataclasses-json==0.6.7",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "executing==2.2.0",
    "ffmpy==0.6.0",
    "filelock==3.18.0",
    "frozenlist==1.7.0",
    "fsspec==2025.5.1",
    "google-crc32c==1.7.1",
    "gradio-client==2.0.0.dev3",
    "greenlet==3.2.3",
    "griffe==1.7.3",
    "groovy==0.1.2",
    "h11==0.16.0",
    "hf-xet==1.1.3",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.34.0",
    "idna==3.10",
    "ifaddr==0.2.0",
    "importlib-metadata==8.7.0",
    "ipython==9.3.0",
    "ipython-pygments-lexers==1.1.1",
    "isodate==0.7.2",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonref==1.1.0",
    "jsonschema==4.24.0",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.4.1",
    "jupyterlab-widgets==3.0.15",
    "langchain==1.2.0",
    "langchain-core==1.2.2",
    "langchain-text-splitters==1.0.0",
    "langgraph-checkpoint==2.1.0",
    "langgraph-prebuilt==1.0.4",
    "langgraph-sdk==0.2.9",
    "lazy-object-proxy==1.11.0",
    "markdown-it-py==3.0.0",
    "markdownify==1.1.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "more-itertools==10.7.0",
    "msal==1.32.3",
    "msal-extensions==1.3.1",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "narwhals==1.42.0",
    "nest-asyncio==1.6.0",
    "numpy==2.3.0",
    "ollama==0.5.1",
    "openapi-core==0.19.5",
    "openapi-schema-validator==0.6.3",
    "openapi-spec-validator==0.7.2",
    "opentelemetry-api==1.39",
    "opentelemetry-sdk==1.39",
    "opentelemetry-semantic-conventions==0.60b0",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "packaging==24.2",
    "pandas==2.3.0",
    "parse==1.20.2",
    "parso==0.8.4",
    "pathable==0.4.4",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "prance==25.4.8.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "protego==0.4.0",
    "protobuf==5.29.5",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pybars4==0.9.13",
    "pycparser==2.22",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.10.1",
    "pydub==0.25.1",
    "pyee==13.0.0",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pylibsrtp==0.12.0",
    "pymeta3==0.5.1",
    "pyopenssl==25.1.0",
    "python-dateutil==2.9.0.post0",
    "python-http-client==3.3.7",
    "pyyaml==6.0.2",
    "readabilipy==0.3.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rpds-py==0.25.1",
    "ruamel-yaml==0.18.14",
    "ruamel-yaml-clib==0.2.12",
    "ruff==0.11.13",
    "safehttpx==0.1.7",
    "scipy==1.15.3",
    "semantic-version==2.10.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.41",
    "sqlite-vec==0.1.6",
    "sse-starlette==2.3.6",
    "stack-data==0.6.3",
    "starlette==0.46.2",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tomlkit==0.13.3",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.16.0",
    "types-requests==2.32.0.20250602",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "werkzeug==3.1.1",
    "widgetsnbextension==4.0.14",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.23.0",
    "googletrans==4.0.2",
    "argparse==1.4.0",
    "utcp==0.1.8",
    "gepa==0.0.7",
    "websockets==15.0.1",
    "python-a2a==0.5.9",
    "google-genai==1.56.0",
    "googlesearch-python==1.3.0",
    "ascii-magic==2.3.0",
    "grandalf==0.8",
    "langchain-google-genai==4.1.2",
    "arxiv==2.2.0",
    "pymupdf==1.26.4",
    "ag2[openai]==0.9.9",
    "pytest==8.4.2",
    "deepeval>=3.7.2",
    "chromadb==1.1.0",
    "peft==0.17.1",
    "datasets==4.2.0",
    "adalflow==1.1.3",
    "mlflow==3.5.1",
    "posthog==5.4.0",
    "json-repair==0.52.4",
    "deepagents==0.2.1",
    "langchain-tavily==0.2.13",
    "streamlit==1.51.0",
    "foundry-local-sdk==0.5.1",
    "googlemaps==4.10.0",
    "polyline>=2.0.4",
    "agent-framework-anthropic==1.0.0b251211",
    "agent-framework-azurefunctions>=1.0.0b251211",
    "onnxruntime ~= 1.21.1",
    "litellm[proxy]==1.80.11",
    "dspy>=3.0.3",
    "grpcio>=1.67.1",
    "uvicorn>=0.31.1",
    "rich>=13.7.1",
    "python-multipart>=0.0.18",
    "fastapi>=0.124.4",
    "fastmcp>=1.0",
    "google-adk>=0.0.1",
    "opik>=1.9.42",
    "opik-optimizer>=2.3.6",
    "prisma>=0.15.0",
    "browser-use>=0.5.5",
    "browserbase>=1.4.0",
    "agno>=2.3.21",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
